+++ idscp2/src/main/java/de/fhg/aisec/ids/idscp2/idscp_core/finite_state_machine/StateWaitForRat.java
@@ -22,7 +22,7 @@ public class StateWaitForRat extends State {
                            Timer verifierHandshakeTimer,
                            Timer proverHandshakeTimer,
                            Timer ratTimer,
-                           int ratTimerDelay,
+                           long ratTimerDelay,
                            DapsDriver dapsDriver) {
